Improved: code to move the action to component(#18)
