+++ src/store/modules/facility/actions.ts
@@ -186,26 +186,6 @@ const actions: ActionTree<FacilityState, RootState> = {
     } catch(err) {
       logger.error('Failed to find the facility locations', err)
     }
-  },
-
-  async deleteFacilityLocation({ dispatch }, payload) {
-    const params = {
-      facilityId: payload.facilityId,
-      locationSeqId: payload.locationSeqId
-    }
-
-    try {
-      const resp = await FacilityService.deleteFacilityLocation(params)
-
-      if(!hasError(resp)) {
-        dispatch('fetchFacilityLocation')
-      } else {
-        throw resp.data
-      }
-    } catch(err) {
-      showToast(translate('Failed to remove facility location'))
-      logger.error('Failed to remove facility location', err)
-    }
   }
 }
